third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-animation-corner-cases.tentative.html


<!DOCTYPE html>
<meta charset="utf-8">
<link rel=author href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/popup-utils.js"></script>

<body>
<style>
.animation { opacity: 0; }
.animation:top-layer { opacity: 1; }
.animation:not(:top-layer) { animation: fade-out 1000s; }
@keyframes fade-out {
  from { opacity: 1; }
  to { opacity: 0; }
}

.animation>div>div { left: 0; }
.animation:not(:top-layer)>div>div { animation: rotate 1000s; color:red;}
@keyframes rotate {
  from { transform: rotate(0); }
  to { transform: rotate(360deg); }
}

[popup] { top: 200px; }
[popup]::backdrop { background-color: rgba(255,0,0,0.2); }
</style>

<script>
function createPopup(t,type) {
  const popup = document.createElement('div');
  popup.popup = 'auto';
  popup.classList = type;
  const div = document.createElement('div');
  const descendent = div.appendChild(document.createElement('div'));
  descendent.appendChild(document.createTextNode("Descendent element"));
  popup.append("This is a popup",div);
  document.body.appendChild(popup);
  t.add_cleanup(() => popup.remove());
  return {popup, descendent};
}
promise_test(async (t) => {
  const {popup, descendent} = createPopup(t,'animation');
  assert_false(isElementVisible(popup));
  assert_equals(descendent.parentElement.parentElement,popup);
  popup.showPopup();
  assert_true(popup.matches(':top-layer'));
  assert_true(isElementVisible(popup));
  assert_equals(popup.getAnimations({subtree: true}).length,0);
  popup.hidePopup();
  const animations = popup.getAnimations({subtree: true});
  assert_equals(animations.length,2,'There should be two animations running');
  assert_false(popup.matches(':top-layer'),'popup should not match :top-layer as soon as hidden');
  assert_true(isElementVisible(popup),'but animations should keep the popup visible');
  assert_true(isElementVisible(descendent),'The descendent should also be visible');
  await waitForRender();
  await waitForRender();
  assert_equals(popup.getAnimations({subtree: true}).length,2,'The animations should still be running');
  assert_true(isElementVisible(popup),'Popup should still be visible due to animation');
  animations.forEach(animation => animation.finish()); // Force the animations to finish
  await waitForRender(); // Wait one frame
  assert_false(popup.matches(':top-layer'),'The popup still shouldn\'t match :top-layer');
  assert_false(isElementVisible(popup),'The popup should now be invisible');
  assert_false(isElementVisible(descendent),'The descendent should also be invisible');
  assert_equals(popup.getAnimations({subtree: true}).length,0);
},'Descendent animations should keep the popup visible until the animation ends');

promise_test(async (t) => {
  const {popup, descendent} = createPopup(t,'');
  assert_equals(popup.classList.length, 0);
  assert_false(isElementVisible(popup));
  popup.showPopup();
  assert_true(popup.matches(':top-layer'));
  assert_true(isElementVisible(popup));
  assert_equals(popup.getAnimations({subtree: true}).length,0);
  // Start an animation on the popup and its descendent.
  popup.animate([{opacity: 1},{opacity: 0}],{duration: 1000000,iterations: 1});
  descendent.animate([{transform: 'rotate(0)'},{transform: 'rotate(360deg)'}],{duration: 1000000,iterations: 1});
  assert_equals(popup.getAnimations({subtree: true}).length,2);
  // Then hide the popup.
  popup.hidePopup();
  assert_false(popup.matches(':top-layer'),'popup should not match :top-layer as soon as hidden');
  assert_equals(popup.getAnimations({subtree: true}).length,2,'animations should still be running');
  await waitForRender();
  assert_equals(popup.getAnimations({subtree: true}).length,2,'animations should still be running');
  assert_false(isElementVisible(popup),'Pre-existing animations should not keep the popup visible');
},'Pre-existing animations should *not* keep the popup visible until the animation ends');
</script>
